+++ db/schema/index.ts
@@ -8,3 +8,4 @@ export * from "./sale";
 export * from "./payment";
 export * from "./cash";
 export * from "./expense";
+export * from "./commission";